Refactor: reorganize exports
