+++ src/App.tsx
@@ -4,12 +4,12 @@ import styled from "@emotion/styled";
 import { inject } from "@vercel/analytics";
 import { useRecoilState } from "recoil";
 
-import ActionButtons from "./components/ActionButtons";
-import HtmlToCanvas from "./components/HtmlToCanvas";
-import InfoAndReport from "./components/InfoAndReport";
-import Popup from "./components/Popup";
-import SavedList from "./components/SavedList";
-import Table from "./components/Table";
+import { ActionButtons } from "./components/Button";
+import { Header } from "./components/Header";
+import { HtmlToCanvas } from "./components/HtmlToCanvas";
+import { Popup } from "./components/Popup";
+import { Table } from "./components/Table";
+import { TimeLogs } from "./components/TimeLogs";
 import { savedWorkTimeState, showKakaoShareState, workTimeState } from "./stores/atoms";
 import { saveLocalStorage } from "./utils/localStorage";
 
@@ -28,14 +28,12 @@ const App: React.FC = () => {
     return (
         <Container>
             <Popup />
-            <InfoAndReport />
+            <Header />
             <main>
-                <TableWrapper>
-                    <Table />
-                </TableWrapper>
+                <Table />
                 <ActionButtons />
                 {showKakaoShare && <HtmlToCanvas />}
-                <SavedList />
+                <TimeLogs />
             </main>
         </Container>
     );
@@ -43,12 +41,6 @@ const App: React.FC = () => {
 
 const Container = styled.div`
     padding: 0 12px;
-`;
-
-const TableWrapper = styled.section`
-    margin-top: 2px;
-    min-width: 300px;
-    overflow-x: scroll;
 `;
 
 export default App;
