+++ src/components/Button/CaptureButton.tsx
