+++ src/components/Header/index.tsx
@@ -1,0 +1,3 @@
+import Header from "./Header";
+
+export { Header };
