+++ src/components/HtmlToCanvas/HtmlToCanvas.tsx
@@ -1,14 +1,14 @@
 import React, { useRef, useState } from "react";
 
-import styled from "@emotion/styled";
 import html2canvas from "html2canvas";
 import { useRecoilState } from "recoil";
 
-import CaptureButton from "./CaptureButton";
-import KakaoShareButton from "./KakaoShare";
-import KakaoSummaryTable from "./SummaryTable";
-import useCloudUploader from "../hooks/useCloudUploader";
-import { summaryTableListState } from "../stores/atoms";
+import * as S from "./HtmlToCanvas.styled";
+import useCloudUploader from "../../hooks/useCloudUploader";
+import { summaryTableListState } from "../../stores/atoms";
+import CaptureButton from "../Button/CaptureButton";
+import KakaoShareButton from "../Button/KakaoShareButton";
+import SummaryTable from "../Table/SummaryTable";
 
 const HtmlToCanvas = () => {
     const tableRef = useRef(null);
@@ -45,37 +45,16 @@ const HtmlToCanvas = () => {
     };
 
     return (
-        <CanvasWrapper>
-            <SummaryContainer>
-                {summaryTableList.length > 0 && <KakaoSummaryTable ref={tableRef} summaryTableList={summaryTableList} />}
-                <ButtonWrapper>
+        <S.CanvasWrapper>
+            <S.SummaryContainer>
+                {summaryTableList.length > 0 && <SummaryTable ref={tableRef} summaryTableList={summaryTableList} />}
+                <S.ButtonWrapper>
                     <CaptureButton onCapture={captureTable} isLoading={isLoading} />
                     <KakaoShareButton imageCheck={imageCheck} imageUrl={capturedImageURL} />
-                </ButtonWrapper>
-            </SummaryContainer>
-        </CanvasWrapper>
+                </S.ButtonWrapper>
+            </S.SummaryContainer>
+        </S.CanvasWrapper>
     );
 };
-
-const CanvasWrapper = styled.section`
-    position: absolute;
-    right: 0;
-    padding-right: 12px;
-`;
-
-const SummaryContainer = styled.div`
-    padding: 12px;
-    border: 1px solid #eaeef4;
-    border-radius: 5px;
-    background-color: #ffffff;
-    z-index: 10;
-`;
-
-const ButtonWrapper = styled.div`
-    display: flex;
-    justify-content: flex-end;
-    margin-top: 12px;
-    gap: 5px;
-`;
 
 export default HtmlToCanvas;
