+++ src/components/HtmlToCanvas/index.tsx
@@ -1,0 +1,3 @@
+import HtmlToCanvas from "./HtmlToCanvas";
+
+export { HtmlToCanvas };
