+++ src/components/Popup/index.tsx
@@ -1,0 +1,3 @@
+import Popup from "./Popup";
+
+export { Popup };
