+++ src/components/Table/index.tsx
@@ -1,0 +1,7 @@
+import Checkbox from "./Checkbox";
+import SummaryTable from "./SummaryTable";
+import Table from "./Table";
+import TableRow from "./TableRow";
+import TextInput from "./TextInput";
+
+export { Checkbox, SummaryTable, Table, TableRow, TextInput };
