+++ src/components/TimeLogs/TimeLogs.styled.ts
@@ -1,0 +1,13 @@
+import styled from "@emotion/styled";
+
+const TimeLogsWrapper = styled.div`
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 5px;
+    padding: 12px;
+    width: 300px;
+    border: 1px solid #eaeef4;
+    border-radius: 5px;
+`;
+
+export { TimeLogsWrapper };
