+++ src/components/TimeLogs/index.tsx
@@ -1,0 +1,3 @@
+import TimeLogs from "./TimeLogs";
+
+export { TimeLogs };
